use histdata for values, fix some issues with cache, other fixes
Implemented ability to configure a " subindex" of a channel, for sensors
such as Netflow which write multiple columns of data for a given
channel.

Changed from table interface to histdata interface. average intervals
are based on the timespan, from no averages, to 5 minutes, hourly and
daily.

fixed issue with reading data from sensors that only have one value and
the JSON result is not an array because of that.

fixed serveral html issues in the query editor partial

changed xsl template to work with histdata interface