Fixes Issue #5, other fixes
Fixes Issue #5 (Unexpected Token <) - this was caused by PRTG returning
a string such as "Not enough data available" as opposed to an XML
response (OR HTTP 400+ (perhaps 404?))

Fixes incorrect code in datasource.js datasource test function. Returns
proper error messages if username or URL are invalid.

Fixes API request by returning a rejected promise with an object which
contains a "message" key, which is then displayed by Grafana instead of
an unknown error or stack trace (related to #5)