Set message value of failed backendRequest method
When querying for a sensor or other object that does not exist, PRTG
returns an HTTP bad request with the body being an XML document
containing an <error> element. The "Message" property of the request
object is set as the value of that element's text. The message returned
from PRTG is then displayed in the Grafana error notification.